Added throw
now have 3 states (idle  thrown and reloading) so the ball has a full animation. it reloads when it falls off screen. but it kinda breaks when it hits the fly, so thats what i want to fix next - also i tried to rename the project after having it open in vscode and it broke the file so I had to redownload the template and copy paste my code into it, sorry if that causes any problems but I think it works fine now